Refactor: Move shared utilities to client module
This commit refactors the Loki codebase by moving shared, lightweight utilities into a new `client` module. This promotes code reuse and reduces duplication across the project. The main `loki/v3` module now imports these utilities from the `client` module, simplifying the overall architecture and dependency management.
