+++ pkg/util/constants/metrics_namespace.go
@@ -1,7 +1,10 @@
 package constants
 
+// Re-export from client module to maintain code reuse
+import "github.com/grafana/loki/client/util"
+
 const (
-	Loki   = "loki"
-	Cortex = "cortex"
-	OTLP   = "otlp"
+	Loki   = util.Loki
+	Cortex = util.Cortex
+	OTLP   = util.OTLP
 )
